requirements.txt
Flask
werkzeug
qrcode
pyzbar
Pillow
uuid